ECA-5801: Corrected expected error message in test. Avoided a NullPointerException
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@25662 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/systemtests/src-test/org/ejbca/core/protocol/cmp/DefaultProfileTest.java
@@ -127,7 +127,9 @@ public class DefaultProfileTest extends CmpTestCase {
     @AfterClass
     public static void afterClass() throws Exception {
         Role role = roleSession.getRole(ADMIN, null, "DefaultProfileTest");
-        roleSession.deleteRoleIdempotent(ADMIN, role.getRoleId());
+        if(role != null) {
+            roleSession.deleteRoleIdempotent(ADMIN, role.getRoleId());    
+        }
     }
     
     public DefaultProfileTest() throws Exception {
